+++ SwiftyButton/SwiftyButton.swift
@@ -12,7 +12,6 @@ import UIKit
 
 public struct SwiftyButtonDefaults {
     public static var buttonColor         = UIColor.cyanColor()
-    public static var highlightedColor    = UIColor.cyanColor()
     public static var shadowColor         = UIColor.blueColor()
     public static var shadowHeight        = CGFloat(3)
     public static var cornerRadius        = CGFloat(3)
@@ -20,6 +19,8 @@ public struct SwiftyButtonDefaults {
     public static var disabledShadowColor = UIColor.darkGrayColor()
     public static var titleFont           = UIFont.systemFontOfSize(20)
     public static var buttonPressDepth    = CGFloat(0.7) // In percentage of shadowHeight
+    
+    public static var highlightedColor: UIColor? = nil
 }
 
 @IBDesignable
@@ -33,7 +34,7 @@ public class SwiftyButton: UIButton {
         }
     }
     
-    @IBInspectable public var highlightedColor: UIColor = SwiftyButtonDefaults.highlightedColor {
+    @IBInspectable public var highlightedColor: UIColor? = SwiftyButtonDefaults.highlightedColor {
         didSet {
             updateBackgroundImages()
         }
@@ -77,47 +78,6 @@ public class SwiftyButton: UIButton {
         }
     }
     
-    // MARK: - Style
-    
-    public enum Style {
-        
-        /**
-         Flat button (no 3D effect) that changes color on touch.
-         Using this style will set shadowHeight = 0 and shadowColor = UIColor.clearColor().
-         
-         - parameter buttonColor:      Button color in the normal state
-         - parameter highlightedColor: Button color in the highlighted state
-         */
-        case Flat(buttonColor: UIColor, highlightedColor: UIColor)
-        
-        /**
-         Pressable (3D) button that moves on press.
-         Using this style will set highlightedColor = buttonColor, and the default shadowHeight will be used.
-         
-         - parameter buttonColor: Button color in the normal state
-         - parameter shadowColor: Shadow color of the button
-         */
-        case Pressable(buttonColor: UIColor, shadowColor: UIColor)
-    }
-    
-    // Note: there is no way to "get" the button style after calling `updateWithStyle`, because all properties
-    // can be changed at any time, making the style concept mutable.
-    public func updateWithStyle(style: Style) {
-        switch style {
-        case .Flat(let buttonColor, let highlightedColor):
-            self.buttonColor      = buttonColor
-            self.highlightedColor = highlightedColor
-            self.shadowColor      = .clearColor()
-            self.shadowHeight     = 0
-            break
-        case .Pressable(let buttonColor, let shadowColor):
-            self.buttonColor      = buttonColor
-            self.highlightedColor = buttonColor
-            self.shadowColor      = shadowColor
-            self.shadowHeight     = SwiftyButtonDefaults.shadowHeight
-        }
-    }
-    
     // MARK: - UIButton
     
     override public var highlighted: Bool {
@@ -128,11 +88,6 @@ public class SwiftyButton: UIButton {
             super.highlighted = newValue
             updateTitleInsets()
         }
-    }
-    
-    public convenience init(style: Style) {
-        self.init()
-        updateWithStyle(style)
     }
     
     public override init(frame: CGRect) {
@@ -165,7 +120,7 @@ public class SwiftyButton: UIButton {
     func updateBackgroundImages() {
         
         let normalImage = SwiftyButton.buttonImageWithColor(buttonColor, shadowHeight: shadowHeight, shadowColor: shadowColor, cornerRadius: cornerRadius)
-        let highlightedImage = SwiftyButton.highlightedButtonImageWithColor(highlightedColor, shadowHeight: shadowHeight, shadowColor: shadowColor, cornerRadius: cornerRadius, buttonPressDepth: buttonPressDepth)
+        let highlightedImage = SwiftyButton.highlightedButtonImageWithColor(highlightedColor ?? buttonColor, shadowHeight: shadowHeight, shadowColor: shadowColor, cornerRadius: cornerRadius, buttonPressDepth: buttonPressDepth)
         let disabledImage = SwiftyButton.buttonImageWithColor(disabledButtonColor, shadowHeight: shadowHeight, shadowColor: disabledShadowColor, cornerRadius: cornerRadius)
         
         setBackgroundImage(normalImage,      forState: .Normal)